+++ adventure/package.json
@@ -17,7 +17,7 @@
     "@types/node": "16",
     "parcel": "^2.7.0",
     "ts-node": "^10.9.1",
-    "typescript": "^4.8.3"
+    "typescript": "~4.7.3"
   },
   "dependencies": {
     "immer": "^9.0.15"
