+++ adventure/src/engine/hexColor.ts
