+++ adventure/src/index.ts
@@ -2,4 +2,4 @@ export { world, convertGame } from "./dsl/dsl";
 export { runGame as runCLIGame } from "./cli-client/run";
 export type { TranslationFile } from "./export-translations/exportTranslations";
 export { exportTranslations } from "./export-translations/exportTranslations";
-export { hexColor } from "./dsl/hexColor";
+export { hexColor } from "./engine/hexColor";
